requirements.txt
ffmpeg-python
openai-whisper
torch
numpy
